move index.js to /src
