+++ tests/test-scrample-write.js
@@ -1,6 +1,6 @@
 require('./scrample');
 
-var utp = require('../index');
+var utp = require('../src');
 var assert = require('assert');
 
 utp.createServer(function(socket) {
