+++ tests/test-write-big.js
@@ -1,4 +1,4 @@
-var utp = require('../index');
+var utp = require('../src');
 var assert = require('assert');
 
 var big = new Buffer(100*1024);
